Fix returns router

+++ server/src/MinhaVida.CodeChallege.VaccinationManagement.API/Controllers/PeopleController.cs
@@ -1,10 +1,11 @@
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Options;
 using MinhaVida.CodeChallege.VaccinationManagement.API.Attributes;
 using MinhaVida.CodeChallege.VaccinationManagement.API.Domain.Commands.People;
 using MinhaVida.CodeChallege.VaccinationManagement.API.Domain.Entities;
 using MinhaVida.CodeChallege.VaccinationManagement.API.Domain.Repositories;
-using System;
+using MinhaVida.CodeChallege.VaccinationManagement.API.Infrastructure;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -15,36 +16,41 @@ namespace MinhaVida.CodeChallege.VaccinationManagement.API.Controllers
     {
         private readonly PeopleRepository peopleRepository;
         private readonly BlobRepository blobRepository;
+        private readonly IOptions<PeopleOptions> peopleOpts;
 
         public PeopleController(PeopleRepository peopleRepository,
-                                BlobRepository blobRepository)
+                                BlobRepository blobRepository,
+                                IOptions<PeopleOptions> peopleOpts)
         {
             this.peopleRepository = peopleRepository;
             this.blobRepository = blobRepository;
+            this.peopleOpts = peopleOpts;
         }
 
         [HttpGet]
-        public async Task<IEnumerable<Person>> Get()
+        public async Task<IActionResult> Get()
         {
             var people = await peopleRepository.GetAsync();
-            return people;
+            return new ObjectResult(people);
         }
 
         [HttpGet("{id}")]
-        public async Task<Person> Get(string id)
+        public async Task<IActionResult> Get(string id)
         {
             Person person = await peopleRepository.GetAsync(id);
-            return person;
+            if (person == null)
+                return new NotFoundResult();
+
+            return new ObjectResult(person);
         }
 
         [HttpPost]
         [ValidateModelState]
         public async Task<IActionResult> Post([FromBody]PeopleCommand command)
         {
-            string defaultUrlImage = "";
-
             var person = new Person(command);
-            person.ChangePhoto(defaultUrlImage);
+            person.ChangePhoto(peopleOpts.Value.DefaultUrlImage);
+
             await peopleRepository.AddAsync(person);
 
             var location = $"{Request.Path}/api/v1/people/{person.Id}";
@@ -85,9 +91,14 @@ namespace MinhaVida.CodeChallege.VaccinationManagement.API.Controllers
 
         [HttpDelete("{id}")]
         [ValidateModelState]
-        public async Task Delete(string id)
+        public async Task<IActionResult> Delete(string id)
         {
+            Person person = await peopleRepository.GetAsync(id);
+            if (person == null)
+                return new NotFoundResult();
+
             await peopleRepository.DeleteAsync(id);
+            return new OkResult();
         }
 
         [HttpPost("{id}/vaccines")]
